refactor(modularization): Phase 4 Step 4-S5 — governance services + views move + G15 role literal compare (#25)
- git mv audit_service + api_key_service → aperag/domains/governance/service/
- Merge views/audit.py + views/api_key.py → aperag/domains/governance/api/routes.py
- G15 canonical落地: user.role == 'admin' 字面量 compare, 不 import Role enum from identity
- AuthenticatedUser(Protocol) 替代 User ORM (G16)
- Legacy 4 shims保 back-compat
- app.py: single governance_router include 替代 2 legacy includes
- OpenAPI byte-stable ✓ (scripts/export_openapi.py --check clean)
- 671 unit tests pass, lint + format clean
- 14th commit on bryce/phase4-impl
